server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
google-generativeai==0.3.2
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-cloud-vision==3.4.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1
pillow==10.1.0
